Refs #124. Moved style basedOn testing to its own module
